updated place_bet, payouts
